Refactor background state and storage boundaries
