+++ app/ts/background/confirmTransactionSimulation.ts
@@ -1,0 +1,127 @@
+import type { EthereumClientService } from '../simulation/services/EthereumClientService.js'
+import { appendTransactionsToInput, getSignedTransactionForSimulation } from '../simulation/services/SimulationModeEthereumClientService.js'
+import type { TokenPriceService } from '../simulation/services/priceEstimator.js'
+import { createSafeExecutionPreSimulationTransaction } from '../safe/safeSimulation.js'
+import type { ConfirmTransactionTransactionSingleVisualization } from '../types/accessRequest.js'
+import type { SafeTransactionSigningRequest } from '../types/safeTypes.js'
+import type { WebsiteCreatedEthereumTransactionOrFailed } from '../types/visualizer-types.js'
+import { last } from '../utils/array.js'
+import { NEW_BLOCK_ABORT, JSON_RPC_ERROR_CODE_INTERNAL_ERROR } from '../utils/constants.js'
+import { decodeEthereumError } from '../utils/errorDecoding.js'
+import { JsonRpcResponseError, reportUnexpectedError, isFailedToFetchError, isNewBlockAbort } from '../utils/errors.js'
+import type { UniqueRequestIdentifier } from '../utils/requests.js'
+import { sendPopupMessageToOpenWindows } from './backgroundUtils.js'
+import { identifyAddress } from './metadataUtils.js'
+import { getSimulationErrorAbis } from './simulationErrorAbi.js'
+import { createSimulationStateWithNonceAndBaseFeeFixing, getCurrentSimulationInput, visualizeSimulatorState } from './simulationUpdating.js'
+import { getTabState } from './storageVariables.js'
+
+let confirmTransactionAbortController = new AbortController()
+export const getConfirmTransactionAbortController = () => confirmTransactionAbortController
+
+export async function refreshConfirmTransactionSimulation(
+	ethereum: EthereumClientService,
+	tokenPriceService: TokenPriceService,
+	activeAddress: bigint,
+	simulationMode: boolean,
+	uniqueRequestIdentifier: UniqueRequestIdentifier,
+	transactionToSimulate: WebsiteCreatedEthereumTransactionOrFailed,
+	safeSigningRequest?: SafeTransactionSigningRequest,
+): Promise<ConfirmTransactionTransactionSingleVisualization | undefined> {
+	const info = {
+		uniqueRequestIdentifier,
+		transactionToSimulate,
+		simulationMode,
+		activeAddress,
+		signerName: (await getTabState(uniqueRequestIdentifier.requestSocket.tabId)).signerName,
+		tabIdOpenedFrom: uniqueRequestIdentifier.requestSocket.tabId,
+	}
+	sendPopupMessageToOpenWindows({ method: 'popup_confirm_transaction_simulation_started' }, 'confirmTransaction')
+	confirmTransactionAbortController.abort(NEW_BLOCK_ABORT)
+	confirmTransactionAbortController = new AbortController()
+	const thisConfirmTransactionAbortController = confirmTransactionAbortController
+	const simulationStartedTimestamp = new Date()
+	const simulationInput = await getCurrentSimulationInput()
+	try {
+		const getNewVisualizedSimulationState = async () => {
+			const preSimulationTransaction = transactionToSimulate.success
+				? safeSigningRequest === undefined
+					? {
+						signedTransaction: getSignedTransactionForSimulation(transactionToSimulate),
+						website: transactionToSimulate.website,
+						created: transactionToSimulate.created,
+						originalRequestParameters: transactionToSimulate.originalRequestParameters,
+						transactionIdentifier: transactionToSimulate.transactionIdentifier
+					}
+					: createSafeExecutionPreSimulationTransaction(transactionToSimulate, safeSigningRequest)
+				: undefined
+			const simulationStateWithNewTransaction = preSimulationTransaction === undefined
+				? simulationInput
+				: appendTransactionsToInput(simulationInput, [preSimulationTransaction], undefined, {}, safeSigningRequest !== undefined)
+			const updatedSimulationState = await createSimulationStateWithNonceAndBaseFeeFixing(simulationStateWithNewTransaction, ethereum)
+			return await visualizeSimulatorState(updatedSimulationState, ethereum, tokenPriceService, thisConfirmTransactionAbortController)
+		}
+		const visualizedSimulatorState = await getNewVisualizedSimulationState()
+		const availableAbis = visualizedSimulatorState.addressBookEntries
+			.map((entry) => 'abi' in entry && entry.abi !== undefined ? entry.abi : undefined)
+			.filter((abiOrUndefined): abiOrUndefined is string => abiOrUndefined !== undefined)
+		if (visualizedSimulatorState.visualizedSimulationState.success === false) {
+			return { statusCode: 'failed', data: {
+				...info,
+				simulationStartedTimestamp,
+				error: { ...visualizedSimulatorState.visualizedSimulationState.jsonRpcError.error, decodedErrorMessage: visualizedSimulatorState.visualizedSimulationState.jsonRpcError.error.message },
+				simulationState: {
+					blockNumber: visualizedSimulatorState.simulationState.blockNumber,
+					simulationConductedTimestamp: new Date(),
+				}
+			} }
+		}
+		const blocks = visualizedSimulatorState.visualizedSimulationState.visualizedBlocks
+		const lastTransaction = last(last(blocks)?.simulatedAndVisualizedTransactions ?? [])
+		return {
+			statusCode: 'success',
+			data: {
+				...info,
+				simulationStartedTimestamp,
+				...visualizedSimulatorState,
+				transactionToSimulate: {
+					...transactionToSimulate,
+					...transactionToSimulate.success ? {
+						transaction: {
+							...transactionToSimulate.transaction,
+							nonce: lastTransaction ? lastTransaction.transaction.nonce : transactionToSimulate.transaction.nonce,
+						} }
+					: { error: {
+						...transactionToSimulate.error,
+						decodedErrorMessage: decodeEthereumError(availableAbis, transactionToSimulate.error).reason
+					} }
+				}
+			}
+		}
+	} catch (error) {
+		if (isNewBlockAbort(error)) return undefined
+		if (isFailedToFetchError(error)) return undefined
+		const isJsonRpcResponseError = error instanceof JsonRpcResponseError
+		if (!isJsonRpcResponseError) await reportUnexpectedError(error, { code: 'confirm_transaction_simulation_failed' })
+
+		const baseError = isJsonRpcResponseError
+			? { code: error.code, message: error.message, data: typeof error.data === 'string' ? error.data : '0x' }
+			: { code: JSON_RPC_ERROR_CODE_INTERNAL_ERROR, message: error instanceof Error ? error.message : 'Unknown simulation error', data: '0x' }
+		const extractToAbi = async (): Promise<readonly string[]> => {
+			const params = transactionToSimulate.originalRequestParameters.params[0]
+			if (!('to' in params)) return []
+			if (params.to === undefined || params.to === null) return []
+			const recipient = params.to
+			return await getSimulationErrorAbis(baseError.data, async () => await identifyAddress(ethereum, thisConfirmTransactionAbortController, recipient))
+		}
+		return { statusCode: 'failed', data: {
+			...info,
+			simulationStartedTimestamp,
+			error: { ...baseError, decodedErrorMessage: isJsonRpcResponseError ? decodeEthereumError(await extractToAbi(), baseError).reason : baseError.message },
+			simulationState: {
+				blockNumber: 0n,
+				simulationConductedTimestamp: new Date()
+			}
+		} }
+	}
+}
