+++ app/ts/background/storageVariables.ts
@@ -4,7 +4,7 @@ import type { PendingChainChangeConfirmationPromise, PendingFetchSimulationStack
 import { type PartialIdsOfOpenedTabs, browserStorageLocalGet, browserStorageLocalGet2, browserStorageLocalRemove, browserStorageLocalSet, browserStorageLocalSet2, getTabStateFromStorage, parseTabStateItems, removeTabStateFromStorage, setTabStateToStorage } from '../utils/storageUtils.js'
 import { CompleteVisualizedSimulation, type EthereumSubscriptionsAndFilters, InterceptorTransactionStack, createPassthroughCompleteVisualizedSimulation } from '../types/visualizer-types.js'
 import { browserStorageLocalSafeParseGet } from '../utils/storageUtils.js'
-import { defaultActiveAddresses, defaultRpcs } from './settings.js'
+import { DEFAULT_ACTIVE_ADDRESSES, DEFAULT_RPCS } from '../config/defaults.js'
 import { type UniqueRequestIdentifier, doesUniqueRequestIdentifiersMatch } from '../utils/requests.js'
 import type { AddressBookEntries, AddressBookEntry, ChainIdWithUniversal } from '../types/addressBookTypes.js'
 import type { SignerName } from '../types/signerTypes.js'
@@ -18,9 +18,21 @@ import type { UnexpectedErrorOccured } from '../types/interceptor-reply-messages
 import { getLargeStateValue, prepareLargeStateWrite, setLargeStateValue, setLargeStateValues } from '../utils/largeStateStore.js'
 import type { InterceptorErrorDiagnostic } from '../types/errorDiagnostics.js'
 import { SafeTransactionStacks } from '../types/safeTypes.js'
+import { createStoredValueRepository } from '../utils/storedValue.js'
 
-export const getIdsOfOpenedTabs = async () => (await browserStorageLocalGet('idsOfOpenedTabs'))?.idsOfOpenedTabs ?? { settingsView: undefined, addressBook: undefined, websiteAccess: undefined, simulationStack: undefined }
-export const setIdsOfOpenedTabs = async (ids: PartialIdsOfOpenedTabs) => await browserStorageLocalSet({ idsOfOpenedTabs: { ...await getIdsOfOpenedTabs(), ...ids } })
+const reportCorruptStoredValue = (label: string) => async (error: unknown) => {
+	console.warn(`${ label } was corrupt:`)
+	console.warn(error)
+}
+
+const idsOfOpenedTabsRepository = createStoredValueRepository({
+	read: async () => (await browserStorageLocalGet('idsOfOpenedTabs')).idsOfOpenedTabs,
+	write: async (idsOfOpenedTabs) => { await browserStorageLocalSet({ idsOfOpenedTabs }) },
+	getDefault: () => ({ settingsView: undefined, addressBook: undefined, websiteAccess: undefined, simulationStack: undefined }),
+})
+
+export const getIdsOfOpenedTabs = idsOfOpenedTabsRepository.get
+export const setIdsOfOpenedTabs = async (ids: PartialIdsOfOpenedTabs) => { await idsOfOpenedTabsRepository.update((previous) => ({ ...previous, ...ids })) }
 
 const pendingTransactionsSemaphore = new Semaphore(1)
 export async function getPendingTransactionsAndMessages(): Promise<readonly PendingTransactionOrSignableMessage[]> {
@@ -78,18 +90,23 @@ export async function setFetchSimulationStackRequestPromise(fetchSimulationStack
 	return await browserStorageLocalSet({ fetchSimulationStackRequestPromise })
 }
 
-const pendingWatchAssetRequestsSemaphore = new Semaphore(1)
-export const getPendingWatchAssetRequests = async () => (await browserStorageLocalGet('pendingWatchAssetRequests'))?.pendingWatchAssetRequests ?? []
+const pendingWatchAssetRequestsRepository = createStoredValueRepository<readonly StoredWatchAssetRequest[]>({
+	read: async () => (await browserStorageLocalGet('pendingWatchAssetRequests')).pendingWatchAssetRequests,
+	write: async (pendingWatchAssetRequests) => { await browserStorageLocalSet({ pendingWatchAssetRequests }) },
+	getDefault: () => [],
+})
+export const getPendingWatchAssetRequests = pendingWatchAssetRequestsRepository.get
 export async function updatePendingWatchAssetRequests(update: (requests: readonly StoredWatchAssetRequest[]) => readonly StoredWatchAssetRequest[]) {
-	return await pendingWatchAssetRequestsSemaphore.execute(async () => {
-		const requests = update(await getPendingWatchAssetRequests())
-		await browserStorageLocalSet({ pendingWatchAssetRequests: requests })
-		return requests
-	})
+	return (await pendingWatchAssetRequestsRepository.update(update)).current
 }
 
-export const getPopupRefreshGeneration = async () => (await browserStorageLocalGet('popupRefreshGeneration'))?.popupRefreshGeneration ?? 0
-export const setPopupRefreshGeneration = async (popupRefreshGeneration: number) => await browserStorageLocalSet({ popupRefreshGeneration })
+const popupRefreshGenerationRepository = createStoredValueRepository({
+	read: async () => (await browserStorageLocalGet('popupRefreshGeneration')).popupRefreshGeneration,
+	write: async (popupRefreshGeneration) => { await browserStorageLocalSet({ popupRefreshGeneration }) },
+	getDefault: () => 0,
+})
+export const getPopupRefreshGeneration = popupRefreshGenerationRepository.get
+export const setPopupRefreshGeneration = popupRefreshGenerationRepository.set
 
 const simulationResultsSemaphore = new Semaphore(1)
 export async function getPopupVisualisationState() {
@@ -116,8 +133,13 @@ export async function updatePopupVisualisationWithCallBack(update: (oldResults: 
 	})
 }
 
-export const setDefaultSignerName = async (signerName: SignerName) => await browserStorageLocalSet({ signerName })
-const getDefaultSignerName = async () => (await browserStorageLocalGet('signerName'))?.signerName ?? 'NoSignerDetected'
+const defaultSignerNameRepository = createStoredValueRepository<SignerName>({
+	read: async () => (await browserStorageLocalGet('signerName')).signerName,
+	write: async (signerName) => { await browserStorageLocalSet({ signerName }) },
+	getDefault: () => 'NoSignerDetected',
+})
+export const setDefaultSignerName = defaultSignerNameRepository.set
+const getDefaultSignerName = defaultSignerNameRepository.get
 
 export async function getTabState(tabId: number) : Promise<TabState> {
 	return await getTabStateFromStorage(tabId) ?? {
@@ -152,63 +174,51 @@ export async function updateTabState(tabId: number, updateFunc: (prevState: TabS
 	})
 }
 
-export const getPendingAccessRequests = async () => (await browserStorageLocalGet('pendingInterceptorAccessRequests'))?.pendingInterceptorAccessRequests ?? []
-const pendingAccessRequestsSemaphore = new Semaphore(1)
+const pendingAccessRequestsRepository = createStoredValueRepository<PendingAccessRequests>({
+	read: async () => (await browserStorageLocalGet('pendingInterceptorAccessRequests')).pendingInterceptorAccessRequests,
+	write: async (pendingInterceptorAccessRequests) => { await browserStorageLocalSet({ pendingInterceptorAccessRequests }) },
+	getDefault: () => [],
+})
+export const getPendingAccessRequests = pendingAccessRequestsRepository.get
 export async function updatePendingAccessRequests(updateFunc: (prevState: PendingAccessRequests) => Promise<PendingAccessRequests>) {
-	return await pendingAccessRequestsSemaphore.execute(async () => {
-		const previous = await getPendingAccessRequests()
-		const pendingAccessRequests = await updateFunc(previous)
-		await browserStorageLocalSet({ pendingInterceptorAccessRequests: pendingAccessRequests })
-		return { previous: previous, current: pendingAccessRequests }
-	})
+	return await pendingAccessRequestsRepository.update(updateFunc)
 }
 
 export async function clearPendingAccessRequests() {
-	return await pendingAccessRequestsSemaphore.execute(async () => {
-		const pending = await getPendingAccessRequests()
-		await browserStorageLocalSet({ pendingInterceptorAccessRequests: [] })
-		return pending
-	})
+	return (await pendingAccessRequestsRepository.update(() => [])).previous
 }
 
 export const saveCurrentTabId = async (tabId: number) => browserStorageLocalSet({ currentTabId: tabId })
 export const getCurrentTabId = async () => (await browserStorageLocalGet('currentTabId'))?.currentTabId ?? undefined
 
-export const setRpcConnectionStatus = async (rpcConnectionStatus: RpcConnectionStatus) => browserStorageLocalSet({ rpcConnectionStatus })
+const rpcConnectionStatusRepository = createStoredValueRepository<RpcConnectionStatus>({
+	read: async () => (await browserStorageLocalGet('rpcConnectionStatus')).rpcConnectionStatus,
+	write: async (rpcConnectionStatus) => { await browserStorageLocalSet({ rpcConnectionStatus }) },
+	getDefault: () => undefined,
+	recover: reportCorruptStoredValue('Connection status'),
+})
+export const setRpcConnectionStatus = rpcConnectionStatusRepository.set
+export const getRpcConnectionStatus = rpcConnectionStatusRepository.get
 
-export async function getRpcConnectionStatus() {
-	try {
-		return (await browserStorageLocalGet('rpcConnectionStatus'))?.rpcConnectionStatus ?? undefined
-	} catch (e) {
-		console.warn('Connection status was corrupt:')
-		console.warn(e)
-		return undefined
-	}
-}
-
-export const getEthereumSubscriptionsAndFilters = async () => (await browserStorageLocalGet('ethereumSubscriptionsAndFilters'))?.ethereumSubscriptionsAndFilters ?? []
-
-const ethereumSubscriptionsSemaphore = new Semaphore(1)
+const ethereumSubscriptionsRepository = createStoredValueRepository<EthereumSubscriptionsAndFilters>({
+	read: async () => (await browserStorageLocalGet('ethereumSubscriptionsAndFilters')).ethereumSubscriptionsAndFilters,
+	write: async (ethereumSubscriptionsAndFilters) => { await browserStorageLocalSet({ ethereumSubscriptionsAndFilters }) },
+	getDefault: () => [],
+})
+export const getEthereumSubscriptionsAndFilters = ethereumSubscriptionsRepository.get
 export async function updateEthereumSubscriptionsAndFilters(updateFunc: (prevState: EthereumSubscriptionsAndFilters) => EthereumSubscriptionsAndFilters) {
-	return await ethereumSubscriptionsSemaphore.execute(async () => {
-		const oldSubscriptions = await getEthereumSubscriptionsAndFilters()
-		const newSubscriptions = updateFunc(oldSubscriptions)
-		await browserStorageLocalSet({ ethereumSubscriptionsAndFilters: newSubscriptions })
-		return { oldSubscriptions, newSubscriptions }
-	})
+	const { previous, current } = await ethereumSubscriptionsRepository.update(updateFunc)
+	return { oldSubscriptions: previous, newSubscriptions: current }
 }
 
-export const setRpcList = async(rpcEntries: RpcEntries) => await browserStorageLocalSet({ rpcEntries })
-
-export async function getRpcList() {
-	try {
-		return (await browserStorageLocalGet('rpcEntries'))?.rpcEntries ?? defaultRpcs
-	} catch(e) {
-		console.warn('Rpc entries were corrupt:')
-		console.warn(e)
-		return defaultRpcs
-	}
-}
+const rpcListRepository = createStoredValueRepository<RpcEntries>({
+	read: async () => (await browserStorageLocalGet('rpcEntries')).rpcEntries,
+	write: async (rpcEntries) => { await browserStorageLocalSet({ rpcEntries }) },
+	getDefault: () => DEFAULT_RPCS,
+	recover: reportCorruptStoredValue('Rpc entries'),
+})
+export const setRpcList = rpcListRepository.set
+export const getRpcList = rpcListRepository.get
 
 export const setInterceptorStartSleepingTimestamp = async(interceptorStartSleepingTimestamp: number) => await browserStorageLocalSet({ interceptorStartSleepingTimestamp })
 
@@ -248,11 +258,11 @@ export async function getUserAddressBookEntries(): Promise<AddressBookEntries> {
 	const { userAddressBookEntriesV3: rawEntries } = await browser.storage.local.get('userAddressBookEntriesV3')
 	const parsedEntries = await browserStorageLocalSafeParseGet('userAddressBookEntriesV3')
 	if (parsedEntries?.userAddressBookEntriesV3 !== undefined) return parsedEntries.userAddressBookEntriesV3
-	if (rawEntries === undefined) return defaultActiveAddresses
+	if (rawEntries === undefined) return DEFAULT_ACTIVE_ADDRESSES
 	console.warn('userAddressBookEntriesV3 was corrupt:')
 	console.warn(rawEntries)
-	await browserStorageLocalSet({ userAddressBookEntriesV3: defaultActiveAddresses })
-	return defaultActiveAddresses
+	await browserStorageLocalSet({ userAddressBookEntriesV3: DEFAULT_ACTIVE_ADDRESSES })
+	return DEFAULT_ACTIVE_ADDRESSES
 }
 export const getUserAddressBookEntriesForChainId = async (chainId: ChainIdWithUniversal) => (await getUserAddressBookEntries()).filter((entry) => entry.chainId === chainId || (entry.chainId === undefined && chainId === 1n) || entry.chainId === 'AllChains')
 export const getUserAddressBookEntriesForChainIdMorePreciseFirst = async (chainId: ChainIdWithUniversal) => {
@@ -278,7 +288,7 @@ export async function updateUserAddressBookEntries(updateFunc: (prevState: Addre
 
 export async function updateUserAddressBookEntriesV2Old(updateFunc: (prevState: AddressBookEntries) => AddressBookEntries) {
 	await userAddressBookEntriesSemaphore.execute(async () => {
-		const entries = (await browserStorageLocalGet('userAddressBookEntriesV2'))?.userAddressBookEntriesV2 ?? defaultActiveAddresses
+		const entries = (await browserStorageLocalGet('userAddressBookEntriesV2')).userAddressBookEntriesV2 ?? DEFAULT_ACTIVE_ADDRESSES
 		return await browserStorageLocalSet({ userAddressBookEntriesV2: updateFunc(entries) })
 	})
 }
