+++ app/ts/background/windows/interceptorAccess.ts
@@ -4,7 +4,8 @@ import type { InterceptorAccessChangeAddress, InterceptorAccessRefresh, Intercep
 import { createScopedKeyedSerialExecutor, Semaphore } from '../../utils/semaphore.js'
 import type { WebsiteTabConnections } from '../../types/user-interface-types.js'
 import { getAssociatedAddresses, persistWebsiteAccessChange, updateWebsiteApprovalAccesses, verifyAccess, withSuppressedUnscopedConnectionEventsForSocket, withSuppressedUnscopedConnectionEventsForSocketAsync } from '../accessManagement.js'
-import { changeActiveAddressAndChain, handleInterceptedRequest, refuseAccess } from '../background.js'
+import { handleInterceptedRequest, refuseAccess } from '../background.js'
+import { changeActiveAddressAndChain } from '../activeSettings.js'
 import { INTERNAL_CHANNEL_NAME, createInternalMessageListener, getHtmlFile, sendPopupMessageToOpenWindows, websiteSocketToString } from '../backgroundUtils.js'
 import { getActiveAddressEntry, getActiveAddresses } from '../metadataUtils.js'
 import { getSettings } from '../settings.js'
