+++ knip.jsonc
@@ -15,6 +15,7 @@
 				"app/ts/popup.ts",
 				"app/ts/settingsView.ts",
 				"app/ts/websiteAccess.ts",
+				"app/ts/watchAsset.ts",
 				"app/inpage/ts/document_start.ts",
 				"app/inpage/ts/inpage.ts",
 				"app/inpage/ts/listenContentScript.ts",
