+++ test/tests/backgroundEthAccountsTestHarness.ts
@@ -130,6 +130,7 @@ export function installBrowserMock({ deferFirstChainChangeRemoval = false, manif
 export async function loadModules() {
 	return {
 		...await import('../../app/ts/background/accessManagement.js'),
+		...await import('../../app/ts/background/activeSettings.js'),
 		...await import('../../app/ts/background/background.js'),
 		...await import('../../app/ts/background/backgroundUtils.js'),
 		...await import('../../app/ts/background/popupMessageHandlers.js'),
