+++ test/tests/backgroundUtilsMessaging.test.ts
@@ -64,6 +64,7 @@ async function loadModules() {
 		...await import('../../app/ts/background/backgroundUtils.js'),
 		...await import('../../app/ts/background/storageVariables.js'),
 		...await import('../../app/ts/background/background.js'),
+		...await import('../../app/ts/background/popupMessageRouting.js'),
 		...await import('../../app/ts/types/interceptor-reply-messages.js'),
 	}
 }
